Fixes #8. Added packages for mappers, reducers and the class that runs the jobs. Organized imports.
