+++ src/main/java/com.griddynamics/customtype/TimestampWritableComparable.java
@@ -1,8 +1,6 @@
-package com.griddynamics;
+package com.griddynamics.customtype;
 
-import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.WritableComparable;
-
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
